Module 3 - Text quest

+++ src/test/java/model/GameStateTest.java
@@ -8,15 +8,15 @@ public class GameStateTest {
     public void statisticsShouldUpdateWinLossAndWinRate() {
         GameState state = new GameState();
         state.setPlayerName("Evgeniy");
-        Assert.assertEquals(0L, (long) state.getTotalWins());
-        Assert.assertEquals(0L, (long) state.getTotalLosses());
+        Assert.assertEquals(0L, state.getTotalWins());
+        Assert.assertEquals(0L, state.getTotalLosses());
         Assert.assertEquals((double) 0.0F, state.getWinRate(), 1.0E-4);
-        Assert.assertEquals(0L, (long) state.getGamesPlayed());
+        Assert.assertEquals(0L, state.getGamesPlayed());
         state.updateStatistics(true);
         state.updateStatistics(false);
-        Assert.assertEquals(1L, (long) state.getTotalWins());
-        Assert.assertEquals(1L, (long) state.getTotalLosses());
-        Assert.assertEquals(2L, (long) state.getGamesPlayed());
+        Assert.assertEquals(1L, state.getTotalWins());
+        Assert.assertEquals(1L, state.getTotalLosses());
+        Assert.assertEquals(2L, state.getGamesPlayed());
         Assert.assertEquals((double) 0.5F, state.getWinRate(), 1.0E-4);
     }
 
@@ -25,10 +25,10 @@ public class GameStateTest {
         GameState state = new GameState();
         state.setPlayerName("Evgeniy");
         state.updateStatistics(true);
-        Assert.assertEquals(1L, (long) state.getTotalWins());
+        Assert.assertEquals(1L, state.getTotalWins());
         state.reset();
-        Assert.assertEquals(1L, (long) state.getTotalWins());
-        Assert.assertEquals(0L, (long) state.getTotalLosses());
+        Assert.assertEquals(1L, state.getTotalWins());
+        Assert.assertEquals(0L, state.getTotalLosses());
         Assert.assertEquals("start", state.getCurrentStepId());
         Assert.assertFalse(state.isGameFinished());
         Assert.assertTrue(state.getChoices().isEmpty());
@@ -40,7 +40,7 @@ public class GameStateTest {
         state.setPlayerName("");
         Assert.assertEquals("Аноним", state.getPlayerName());
         state.updateStatistics(true);
-        Assert.assertEquals(1L, (long) state.getTotalWins());
-        Assert.assertEquals(0L, (long) state.getTotalLosses());
+        Assert.assertEquals(1L, state.getTotalWins());
+        Assert.assertEquals(0L, state.getTotalLosses());
     }
 }
